Fixed Defect

+++ Calculator/square.py
@@ -1,4 +1,4 @@
 def square(a):
-    a = int(a)
+    a = float(a)
     b = a*a
     return b
